Apply translate to Shape
