+++ shoes-core/lib/shoes/shape.rb
@@ -1,10 +1,11 @@
 class Shoes
   class Shape
     include Common::UIElement
+    include Common::Clickable
     include Common::Fill
     include Common::Stroke
     include Common::Style
-    include Common::Clickable
+    include Common::Translate
 
     attr_reader :blk, :x, :y, :left_bound, :top_bound, :right_bound, :bottom_bound
