+++ shoes-swt/lib/shoes/swt.rb
@@ -74,6 +74,7 @@ class Shoes
       require 'shoes/swt/line_painter.rb'
       require 'shoes/swt/oval_painter.rb'
       require 'shoes/swt/rect_painter.rb'
+      require 'shoes/swt/shape_painter.rb'
       require 'shoes/swt/swt_button'
       require 'shoes/swt/check_button'
       require 'shoes/swt/radio_group'
