+++ shoes-swt/lib/shoes/swt/shape_painter.rb
@@ -1,0 +1,13 @@
+class Shoes
+  module Swt
+    class ShapePainter < Common::Painter
+      def fill(gc)
+        gc.fill_path(@obj.element)
+      end
+
+      def draw(gc)
+        gc.draw_path(@obj.element)
+      end
+    end
+  end
+end
